requirements.txt
httpx
beautifulsoup4
phidata
groq
openai
requests
fastapi
uvicorn